Put mutation tests beside their sources
